Add Changes to Existing Code
The original code is in the production and works fine. The new senario is to add a type for each employee. One way is to start modifying PersonModel, EmployeeModel, Accounts and Program.cs. Which is against OCP.
The next step is to undo the changes and find better way without modifiying or make very small changes to the existing code.